fix in framerate calculations
the number of cue alterations cannot be a float

+++ opcBridge.py
@@ -190,7 +190,7 @@ def multiCommand(commands):
         for i in range(abs(appends)):
             queueList.append({})
     for c in commands:
-        commandAlterations = c[2] * frameRate
+        commandAlterations = int(c[2] * frameRate)
         for i in range(c[0][0], c[0][1]):
             start = pixels[i]
             bridgeGenerator = bridgeValues(commandAlterations, start, c[1])
